requirements.txt
aiohttp==1.0.5
discord.py==0.16.0
requests==2.10.0
twilio==5.6.0
Flask==0.12
gunicorn==19.6.0